+++ src/entities/card.py
@@ -27,12 +27,9 @@ class Card:
 
         return str(self.rank) + str(self.suit)
 
-    # def __repr__(self):
-    #    """Representation of the card"""
-
-    #    return self.__str__()
-
     def __lt__(self, other):
+        """Less than comparison for cards"""
+
         return self.rank < other.rank
 
     def get_card(self):